+++ flink-runtime/src/main/java/org/apache/flink/runtime/rest/handler/job/checkpoints/CheckpointStatsCache.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.runtime.rest.handler.legacy.checkpoints;
+package org.apache.flink.runtime.rest.handler.job.checkpoints;
 
 import org.apache.flink.runtime.checkpoint.AbstractCheckpointStats;
